refactor(simulation): move next question to dedicated class
